software/dummy: Move empty impls out
